WIP for additional operations and improving integration testing.

+++ src/InfluxMigrations/InfluxMigrations.Operations/Bucket/RemoveMemberFromBucket.cs
@@ -7,8 +7,8 @@ public class RemoveMemberFromBucket : IMigrationOperation
 {
     private IOperationExecutionContext _context;
 
-    public InfluxRuntimeIdResolver User { get; private set; }
-    public InfluxRuntimeIdResolver Bucket { get; private set; }
+    public IInfluxRuntimeResolver User { get; private set; }
+    public IInfluxRuntimeResolver Bucket { get; private set; }
 
     public RemoveMemberFromBucket(IOperationExecutionContext context)
     {
@@ -29,15 +29,31 @@ public class RemoveMemberFromBucket : IMigrationOperation
         try
         {
             var userId = await User.GetAsync(_context);
-            var bucketId = await User.GetAsync(_context);
+            if (string.IsNullOrEmpty(userId))
+            {
+                return OperationResults.ExecuteFailed("Cannot remove Membership, cannot find User.");
+            }
+            
+            var bucketId = await Bucket.GetAsync(_context);
+            if (string.IsNullOrEmpty(bucketId))
+            {
+                return OperationResults.ExecuteFailed($"Cannot remove Membership, cannot find Bucket.");
+            }
 
+            var membership = await _context.Influx.GetBucketsApi().GetMembersAsync(bucketId);
+
+            if (membership.All(x => x.Id != userId))
+            {
+                return OperationResults.ExecuteFailed($"Could not find ResourceMember for User in Bucket.");
+            }
+            
             await _context.Influx.GetBucketsApi().DeleteMemberAsync(userId, bucketId);
 
             return OperationResults.ExecuteSuccess();
         }
         catch (Exception x)
         {
-            return OperationResults.ExecutionFailed(x);
+            return OperationResults.ExecuteFailed(x);
         }
     }
 
@@ -51,7 +67,7 @@ public class RemoveMemberFromBucket : IMigrationOperation
         try
         {
             var userId = await User.GetAsync(_context);
-            var bucketId = await User.GetAsync(_context);
+            var bucketId = await Bucket.GetAsync(_context);
 
             await _context.Influx.GetBucketsApi().AddMemberAsync(userId, bucketId);
 
@@ -66,10 +82,10 @@ public class RemoveMemberFromBucket : IMigrationOperation
 
 public class RemoveMemberFromBucketBuilder : IMigrationOperationBuilder
 {
-    public string UserName { get; init; }
-    public string UserId { get; init; }
-    public string BucketName { get; init; }
-    public string BucketId { get; init; }
+    public string UserName { get; private set; }
+    public string UserId { get; private set; }
+    public string BucketName { get; private set; }
+    public string BucketId { get; private set; }
 
     public IMigrationOperation Build(IOperationExecutionContext context)
     {
@@ -82,5 +98,29 @@ public class RemoveMemberFromBucketBuilder : IMigrationOperationBuilder
                 .WithName(StringResolvable.Parse(UserName))
                 .WithId(StringResolvable.Parse(UserId));
         });
+    }
+
+    public RemoveMemberFromBucketBuilder WithBucketName(string s)
+    {
+        BucketName = s;
+        return this;
+    }
+
+    public RemoveMemberFromBucketBuilder WithBucketId(string s)
+    {
+        BucketId = s;
+        return this;
+    }
+
+    public RemoveMemberFromBucketBuilder WithUserName(string s)
+    {
+        UserName = s;
+        return this;
+    }
+
+    public RemoveMemberFromBucketBuilder WithUserId(string s)
+    {
+        UserId = s;
+        return this;
     }
 }
